+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# REPLACE-GLOBAL-WITH-NEXT-VERSION
+- PT-11294 - Implemented error handling for Smart Payment Buttons
+
 # 1.5.0
 - NEXT-8322 - Shopware 6.2 compatibility
 - PT-10654 - Activate and set PayPal as default for the selected Saleschannel in the settings module